refactor: stoneEntity와 Stone을 변환하여 MainActivity에게 제공
